rename awsrdsiam to awsiamrds to match v1
