+++ docs/cli/server_dbcrypt_rotate.md
@@ -25,7 +25,7 @@ The connection URL for the Postgres database.
 
 |             |                                        |
 | ----------- | -------------------------------------- |
-| Type        | <code>enum[password\|awsrdsiam]</code> |
+| Type        | <code>enum[password\|awsiamrds]</code> |
 | Environment | <code>$CODER_PG_CONNECTION_AUTH</code> |
 | Default     | <code>password</code>                  |
