+++ enterprise/cli/server_dbcrypt.go
@@ -12,7 +12,7 @@ import (
 	"cdr.dev/slog/sloggers/sloghuman"
 	"github.com/coder/coder/v2/cli"
 	"github.com/coder/coder/v2/cli/cliui"
-	"github.com/coder/coder/v2/coderd/database/awsrdsiam"
+	"github.com/coder/coder/v2/coderd/database/awsiamrds"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/enterprise/dbcrypt"
 	"github.com/coder/serpent"
@@ -91,8 +91,8 @@ func (*RootCmd) dbcryptRotateCmd() *serpent.Command {
 			}
 
 			sqlDriver := "postgres"
-			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSRDSIAM {
-				sqlDriver, err = awsrdsiam.Register(inv.Context(), sqlDriver)
+			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSIAMRDS {
+				sqlDriver, err = awsiamrds.Register(inv.Context(), sqlDriver)
 				if err != nil {
 					return xerrors.Errorf("register aws rds iam auth: %w", err)
 				}
@@ -156,8 +156,8 @@ func (*RootCmd) dbcryptDecryptCmd() *serpent.Command {
 			}
 
 			sqlDriver := "postgres"
-			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSRDSIAM {
-				sqlDriver, err = awsrdsiam.Register(inv.Context(), sqlDriver)
+			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSIAMRDS {
+				sqlDriver, err = awsiamrds.Register(inv.Context(), sqlDriver)
 				if err != nil {
 					return xerrors.Errorf("register aws rds iam auth: %w", err)
 				}
@@ -212,8 +212,8 @@ Are you sure you want to continue?`
 
 			var err error
 			sqlDriver := "postgres"
-			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSRDSIAM {
-				sqlDriver, err = awsrdsiam.Register(inv.Context(), sqlDriver)
+			if codersdk.PostgresAuth(flags.PostgresAuth) == codersdk.PostgresAuthAWSIAMRDS {
+				sqlDriver, err = awsiamrds.Register(inv.Context(), sqlDriver)
 				if err != nil {
 					return xerrors.Errorf("register aws rds iam auth: %w", err)
 				}
